requirements.txt
openai~=1.63.0
pyyaml~=6.0.2
prompt-toolkit~=3.0.50
ollama~=0.4.7